Rename views and view states
