+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/analytics/AnalyticsViewState.kt
@@ -1,7 +1,7 @@
 package com.woocommerce.android.ui.analytics
 
 import com.woocommerce.android.ui.analytics.daterangeselector.AnalyticsDateRangeSelectorViewState
-import com.woocommerce.android.ui.analytics.informationcard.AnalyticsCardInformationViewState
+import com.woocommerce.android.ui.analytics.informationcard.AnalyticsInformationViewState
 
 data class AnalyticsViewState(val analyticsDateRangeSelectorState: AnalyticsDateRangeSelectorViewState,
-                              val revenueCardState: AnalyticsCardInformationViewState.DataViewState)
+                              val revenueState: AnalyticsInformationViewState.DataViewState)
