+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/analytics/informationcard/AnalyticsInformationViewState.kt
@@ -1,0 +1,9 @@
+package com.woocommerce.android.ui.analytics.informationcard
+
+sealed class AnalyticsInformationViewState {
+    object HiddenViewState : AnalyticsInformationViewState()
+    data class DataViewState(val title: String,
+                             val totalValues: AnalyticsInformationSectionViewState,
+                             val netValues: AnalyticsInformationSectionViewState)
+        : AnalyticsInformationViewState()
+}
